tg_requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b8
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.1
altgraph==0.17.4
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
boto3==1.33.8
botocore==1.33.8
certifi==2023.11.17
charset-normalizer==3.3.2
darkdetect==0.8.0
frozenlist==1.4.0
greenlet==3.0.1
idna==3.6
importlib-metadata==6.8.0
importlib-resources==6.1.1
jmespath==1.0.1
magic-filter==1.0.12
Mako==1.3.0
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.2
Paginator==0.5.1
pefile==2023.2.7
Pillow==9.5.0
pydantic==2.0.3
pydantic_core==2.3.0
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
python-dateutil==2.8.2
pytz==2023.3.post1
s3transfer==0.8.2
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==1.26.18
yarl==1.9.3
zipp==3.17.0
